requirements.txt
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crcmod==1.7
cryptography==42.0.3
Django==4.2.10
idna==3.6
jmespath==0.10.0
oss2==2.15.0
pillow==10.2.0
pycparser==2.21
pycryptodome==3.20.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
